lib: aws_iot: honor CONFIG_AWS_IOT_CLIENT_ID_MAX_LEN
This corrects the checks for the allowed maximum client id string
length to be equal the `CONFIG_AWS_IOT_CLIENT_ID_MAX_LEN` value.

Previously the client id string was rejected if it had the maximum
allowed length.

Also corrects the invalid `AWS_TOPIC_LEN`. Previously the library
would not populate the client id into the shadow topic properyl
due to mismatch in expected length for the topics after snprintf.

Co-Authored-by: Jan Tore Guggedal <JanTore.Guggedal@nordicsemi.no>
Co-Authored-by: Simen Røstad <Simen.Rostad@nordicsemi.no>
Signed-off-by: Markus Tacker <Markus.Tacker@NordicSemi.no>