openthread: zigbee: lib: Fix RAM power down functionality
The ram_pwrdn functionality was not disabling all unused RAM
when only part of the block was to be disabled.
There was no need to adjust retenetion settings
as it is disabled by default.
The Kconfig documentation has been adjusted to clarify
that the RAM in the System On mode is powered off.

Signed-off-by: Piotr Szkotak <piotr.szkotak@nordicsemi.no>