doc: add table-from-rows Sphinx extension
Add a new Sphinx extension to ease the creation of sample requirement
tables. To add a table one can add a directive like:

```
.. table-from-rows: /includes/blabla.txt
   :header: header_row
   :rows: row1, row2, ...
```

The input file `/includes/blabla.txt` must have the following format:

```
.. header_row

| col1 | col2 | col3 |

.. row1

| data1 | data2 | data3 |

.. row2

| data1 | data2 | data3 |

.. and so on and so forth...
```

Basic error checks are done, like the header and all rows match in
the number of columns, header and rows referenced in the directive
exists, etc.

Spaces are not relevant but `|` must be used to separate columns, there
must be one at the beginning and at the end of the line. The adjustment
of column sizes if made by the extension before sending for Sphinx
processing. If a section has more that one row they span.

JIRA: NCSDK-6297

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>