drivers: gps: nrf9160_gps: Change log level
If the GPS is restarted during an active GPS search
the setting of the GPS retry parameter will fail. The
GPS-service is then reinitialized before the
aforementioned parameter is successfully set. This is
expected behaviour and the error message propogated
upon the parameter set attempt should be treated as
a debug output rather than a warning.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>