tests: nrfx_integration_test: Fix test case configurations
Provide configurations for 3 test cases:
- without BLE
- with BLE using the SoftDevice LL
- with BLE using the Zephyr SW split LL

The last two cases are accordingly filtered out if the given LL is not
available for a particular target (e.g. for nRF9160, which does not
have the BLE capable radio).

In configuration with CLOCK_CONTROL_MPSL enabled, do not compile
the nrfx_clock driver, as in such case a mocked version of this
driver needs to be used.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>